Storybook : Add Crop-Details Component
