+++ package.json
@@ -30,7 +30,7 @@
     "@fortawesome/free-brands-svg-icons": "^5.12.0",
     "@fortawesome/free-solid-svg-icons": "^5.11.2",
     "@fortawesome/react-fontawesome": "^0.1.7",
-    "@storybook/addon-actions": "^5.3.14",
+    "@storybook/addon-actions": "^5.3.17",
     "@storybook/addon-knobs": "^5.3.14",
     "@storybook/react": "^5.2.6",
     "@storybook/theming": "^5.2.8",
