--- src/client/components/Crop1/AddCrop1/addcrop.css1
@@ -1,64 +1,0 @@
-.accordion {
-  display: flexbox;
-  justify-content: center;
-}
-.accordion .accordion-item:first-child:last-child .title {
-  background: #709d68;
-  border: 1px solid rgba(255, 255, 255, 0.2);
-  padding: 16px;
-  color: #ffffff;
-  font-family: Proxima Nova, sans-serif;
-  text-transform: uppercase;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 62.5%;
-  line-height: 10px;
-  /* button-shadow */
-  box-shadow: -3px -3px 10px rgba(255, 255, 255, 0.8), 3px 3px 15px #c9cbd8;
-  border-radius: 40px;
-}
-
-.accordion .accordion-item .title {
-  width: 600px;
-  text-align: start;
-  background: #709d68;
-  border: 1px solid rgba(255, 255, 255, 0.2);
-  padding: 10px;
-  margin: 2px;
-  color: #ffffff;
-  font-family: Proxima Nova, sans-serif;
-  text-transform: uppercase;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 62.5%;
-  line-height: 10px;
-  /* button-shadow */
-  box-shadow: -3px -3px 10px rgba(255, 255, 255, 0.8), 3px 3px 15px #c9cbd8;
-  border-radius: 40px;
-}
-
-.accordion .accordion-item .panel {
-  max-height: 0;
-  transition: max-height 0.15s ease-out;
-  overflow: hidden;
-}
-.accordion .accordion-item.active .title {
-  font-family: Proxima Nova, sans-serif;
-  font-style: normal;
-  font-weight: bold;
-  text-transform: uppercase;
-  font-size: 62.5%;
-  line-height: 10px;
-  /* button-shadow */
-  box-shadow: -3px -3px 10px rgba(255, 255, 255, 0.8), 3px 3px 15px #c9cbd8;
-  border-radius: 40px;
-}
-
-.accordion .accordion-item.active .panel {
-  max-height: 400px;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 62.5%;
-  line-height: 4px;
-  padding-left: 10px;
-}
