--- src/client/components/Crop1/AddCrop1/addcrop.stories.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import AddCrop from './addcrop';
-
-export default {
-  title: 'AddCrop',
-};
-
-export const AddCropStories = () => <AddCrop />;
