--- src/client/components/Crop1/Button1/Button.css1
@@ -1,23 +1,0 @@
-.Btn {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  text-align: center;
-  float: left;
-  background-color: var(--interaction-color);
-  color: white;
-  float: right;
-
-  width: 200px;
-  height: 32px;
-
-  font-family: Proxima Nova, sans-serif;
-  font-style: normal;
-  font-weight: bold;
-  text-transform: uppercase;
-
-  /* Button shadow */
-  box-shadow: -3px -3px 10px rgba(255, 255, 255, 0.8), 3px 3px 15px #c9cbd8;
-  border-radius: 40px;
-  border: none;
-}
