--- src/client/components/Crop1/Button1/Button.stories.js1
@@ -1,7 +1,0 @@
-import React from 'react';
-import Button from './button';
-
-export default {
-  title: 'Button',
-};
-export const button = () => <Button placeholder="Save Crop Details" />;
