--- src/client/components/Crop1/FormField1/FormField.css1
@@ -1,9 +1,0 @@
-.form-container {
-  margin-top: 0px;
-  position: relative;
-  height: 250px;
-  font-size: 30px;
-  border-radius: 300px;
-  align-items: left;
-  width: 600px;
-}
