--- src/client/components/Crop1/Input1/InputDate.css1
@@ -1,21 +1,0 @@
-.input-date {
-  width: 600px;
-  height: 32px;
-  background: var(--white);
-  border: 1px solid rgba(255, 255, 255, 0.2);
-  box-sizing: border-box;
-  padding: 16px;
-
-  color: #ffffff;
-  font-family: Proxima Nova, sans-serif;
-  font-style: normal;
-  font-size: 62.5%;
-  line-height: 24px;
-  /* button-shadow */
-  box-shadow: -3px -3px 10px rgba(255, 255, 255, 0.8), 3px 3px 15px #c9cbd8;
-  border-radius: 40px;
-}
-.input-date:hover {
-  background-color: #ccc;
-  cursor: pointer;
-}
