--- src/client/components/Crop1/Input1/InputText.css1
@@ -1,25 +1,0 @@
-.input-text {
-  display: flex;
-  align-items: center;
-
-  width: 600px;
-  height: 32px;
-  background: var(--white);
-  border: 1px solid rgba(255, 255, 255, 0.2);
-  box-sizing: border-box;
-  padding: 16px;
-
-  color: #ffffff;
-  font-family: Proxima Nova, sans-serif;
-  font-style: normal;
-  font-size: 62.5%;
-  line-height: 1.5em;
-  /* button-shadow */
-  box-shadow: -3px -3px 10px rgba(255, 255, 255, 0.8), 3px 3px 15px #c9cbd8;
-  border-radius: 40px;
-}
-
-.input-text:hover {
-  background-color: #ccc;
-  cursor: pointer;
-}
