+++ src/client/components/crop/add-crop/add-crop.css
@@ -1,0 +1,27 @@
+.accordion .accordion-item .title {
+  display: flex;
+  background: var(--green);
+  color: rgb(255, 255, 255);
+  border: var(--border-inputs);
+  border-radius: var(--border-radius-high);
+  font-size: 1em;
+  margin: 0.6em 0 0.5em;
+  outline: none;
+  padding: 0.5em 1.2em 0.9em;
+  width: 39em;
+}
+
+.accordion .accordion-item .panel {
+  max-height: 0;
+  transition: max-height 0.15s ease-out;
+  overflow: hidden;
+}
+
+.accordion .accordion-item.active .panel {
+  max-height: 22em;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 1em;
+  line-height: 0.2em;
+  padding-left: 0em;
+}
