+++ src/client/components/crop/add-crop/add-crop.stories.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import AddCrop from './add-crop';
+
+export default {
+  title: 'AddCrop',
+};
+
+export const AddCropPage = () => <AddCrop />;
