+++ src/client/components/crop/form-field/form-field.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import FormField from './formfield';
+import FormField from './form-field';
 
 export default { title: 'FormField' };
 export const Form = () => <FormField />;
