+++ src/client/components/crop/form-title/form-title.css
@@ -6,4 +6,5 @@
   line-height: 42px;
   text-transform: uppercase;
   color: #3b4349;
+  padding-left: 1em;
 }
