+++ src/client/components/crop/form-title/form-title.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import './formtitle.css';
+import './form-title.css';
 
 export default function FormTitle({ title }) {
   return <div className="form-title">{title}</div>;
