+++ src/client/components/crop/input/input-date.css
@@ -1,0 +1,17 @@
+.input-date {
+  display: flex;
+  background: var(--light-grey);
+  color: rgba(59, 67, 73, 0.4);
+  border: var(--border-inputs);
+  border-radius: var(--border-radius-high);
+  box-shadow: var(--debossed-shadow-high);
+  font-size: 1em;
+  margin: 0.6em 0 0.5em;
+  outline: none;
+  padding: 0.5em 1.2em 0.9em;
+  width: 39em;
+}
+.input-date:hover {
+  background-color: transparent;
+  cursor: pointer;
+}
